Update .gitignore to include all JSON files and remove unnecessary files
