+++ .gitignore
@@ -1,4 +1,4 @@
 .env
-function_changes.json
+**.json
 path
 .DS_Store
